refactor: add payment module to better support adding other payment processors
improvement: checks if invoice is already paid
